requirements.txt
openai
selenium
python-dotenv